Version appendix for all names and symbols
this should patch all symbols, constants, module name and a bunch
of other names so that they include a version number (14).

now, the module can be used concurrently with batman v15.
Note that you can not use both protocols on top of the same
interface. However, with a trick you can create a similar effect:
- create two veth pairs (say veth0,veth1 and veth2,veth3)
- bridge veth0 and veth2 and the interface you want to run
  both protocols on
- set up batman v14 on veth1
- set up batman v15 on veth3

Note that the main purpose is not to run both protocols within
the same network, but rather use both protocols on one machine,
using a single kernel instance.